Fix and refactor tests  - moved MAC tests to separate module  - fixed to check for errors
