+++ test/heartbeats/heartbeats.js
@@ -6,6 +6,7 @@ var request = require('supertest'),
   db = require('./../../lib/db'),
   rest = proxyquire('./../../lib/heartbeats', {'./../auth': auth, './../db': db}),
   bodyParser = require('body-parser'),
+  config = require('./../../lib/config'),
   factory = require('./../setup');
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({extended: true}));
@@ -16,6 +17,8 @@ app.set('sockets', {connected : {}, }); //mock sockets
 
 describe('Heartbeats Tests', function() {
   beforeEach(function (done) {
+    config.signatureVerification = false;
+
     db.sequelize.sync({force: true}).then(function (err) {
       factory.create('testUser', function (err, u) {
         if (err) {
@@ -78,47 +81,6 @@ describe('Heartbeats Tests', function() {
                "battery": {"batteryHealth": {"$iLike": "%"}, "batteryPercentage": 100.0, "temperature": 97, "status": 100, "plugged": 1, "date": "2016-08-19T23:11:15.123Z"}})
         .expect(400)
         .end(function (err, res) { should.not.exist(err); done(); });
-    });
-
-    it('cannot json sql inject simid', function (done) {
-      request(app).post('/heartbeats')
-        .send({"mac": {"length": 5}, "imei":"foo","simid":{"$iLike": "%"}})
-        .expect(403)
-        .end(function (err, res) { done(); });
-    });
-
-    it('cannot json sql inject imei', function (done) {
-      request(app).post('/heartbeats')
-        .send({"mac": {"length": 200}, "imei":{"$iLike": "1%"},"simid":"bar"})
-        .set('Authorization', 'Bearer sdfdf')
-        .expect(403)
-        .end(function (err, res) { done(); });
-    });
-
-    it('cannot json sql inject imei and simid', function (done) {
-      request(app).post('/heartbeats')
-        .send({"mac": {"length": 200}, "imei":{"$iLike": "1%"},"simid":{"$iLike": "%"}})
-        .set('Authorization', 'Bearer sdfdf')
-        .expect(403)
-        .end(function (err, res) { done(); });
-    });
-
-    it('cannot perform buffer memory exhaustion using fake arraylike', function (done) {
-      request(app).post('/heartbeats')
-        .send({"mac": {"length": 1000000000}, "imei":"990000862471854","simid":"8991101200003204510"})
-        .set('Authorization', 'Bearer sdfdf')
-        .expect(403)
-        .end(function (err, res) { done(); });
-    });
-
-    it('cannot perform buffer memory exhaustion using long string', function (done) {
-      var mac = Buffer(1025);
-      mac.fill('X');
-      request(app).post('/heartbeats')
-        .send({"mac": mac.toString(), "imei":"990000862471854","simid":"8991101200003204510"})
-        .set('Authorization', 'Bearer sdfdf')
-        .expect(403)
-        .end(function (err, res) { done(); });
     });
   });
